Added Export Attendance List
For the sake of testing against various forms of data, I've added the ability to choose whether you want the dates pre-populated, or blank. Added a button to export the list which would include the student's entire attendance history to be loaded back in, or used outside of the application.