Fixed New Day Logic and Import Functionality
Also created a clear local storage button, for testing to reduce the need to jump between the Application and Component tabs in devTools. Added ability to press any key to cancel removal confirmation. Attempted to convert my remove buttons to SVGs but rolled it back and commented out due to latency when swapping order.